Begin implementation of x86_64 support
libc is already done pretty much, but there's _a_ _lot_ that's still
left to be done within the kernel (instruction emulation in particular
could become a real nightmare)

+++ src/kernel/modules/vga/vga.c
@@ -639,7 +639,7 @@ VGA_Ioctl(VGA *__restrict self,
  } break;
 
  case VGA_SETMODE_DEFAULT:
-  switch ((unsigned int)arg) {
+  switch ((unsigned int)(uintptr_t)arg) {
 
   case VGA_DEFAULT_MODE_TXT80X25_16:
    VGA_SetMode(self,&vga_biosmode);
@@ -675,7 +675,7 @@ VGA_Ioctl(VGA *__restrict self,
   break;
 
  case VGA_SETPAL_DEFAULT:
-  switch ((unsigned int)arg) {
+  switch ((unsigned int)(uintptr_t)arg) {
 
   case VGA_DEFAULT_PAL_TXT16:
    VGA_SetPalette(self,&vga_biospal);
